+++ crates/knipsa-ffi/include/knipsa.h
@@ -10,8 +10,8 @@
 
 /** Semantic version of the headers and matching knipsa library. */
 #define KNIPSA_VERSION_MAJOR 0
-#define KNIPSA_VERSION_MINOR 2
-#define KNIPSA_VERSION_PATCH 1
+#define KNIPSA_VERSION_MINOR 3
+#define KNIPSA_VERSION_PATCH 0
 
 #ifdef __cplusplus
 extern "C" {